Fixed a bunch of compiler warnings

+++ src/packets/chat_message.c
@@ -77,7 +77,7 @@ int chat_message_new(TALLOC_CTX *ctx, const struct player *player, const struct 
 	}
 	
 	packet->type = PACKET_TYPE_CHAT_MESSAGE;
-	packet->len = PACKET_HEADER_SIZE + strlen(message) + binary_writer_7bit_len(strlen(message));
+	packet->len = PACKET_HEADER_SIZE + (uint16_t)strlen(message) + binary_writer_7bit_len(strlen(message));
 
 	chat_message->id = player->id;
 	chat_message->colour = colour;
